+++ lib/components/widget_helpers.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/widgets.dart';
+
+class CenterText extends StatelessWidget {
+  const CenterText(this.text, {Key? key}) : super(key: key);
+
+  final String text;
+
+  @override
+  Widget build(BuildContext context) {
+    return Center(child: Text(text));
+  }
+}