+++ lib/options/preferences.dart
@@ -1,0 +1,10 @@
+/// Global instance of [UserPreference].
+///
+/// There are usually only one user preferences instance in the whole app.
+/// You can use it as the default value of the controller widgets.
+///
+const preference = UserPreference();
+
+class UserPreference {
+  const UserPreference();
+}